MAT-7241 metadata validation moved to ModelValidator
